v2python/table_tool: add --round_inputs
This is a temporary workaround for bug
https://github.com/ROCm/triton-internal/issues/177

The tuning process always use regular inputs, but kernels that are only
faulty for irregular shapes may slip into the tuning database.

In such case, we can rerun the tune_flash with irregular shapes, and
pass the json to table_tool with --round_inputs, to patch corresponding
entries (always regular).