Make the LazyTensor disposal explicit.
Should not make it automatic in destructor, it is possible that dq_acc
is allocated but not used. In this case the dq is filled by the fallback
backend and should not be overriden by dq_acc.

Current AITER ASM bwd kernel does not have this problem, but make it
explicit help avoiding this problem in the future.

Another thing is to track the actual backend used in Op context.

A potential TODO is to change the API of LazyTensor to separate copying
back and free, but in term of performance we should not, being defensive
before actual allocation of dq_acc (or any other intermediate Tensor) is
more preferable. Separation of free from disposal only hides the
performance problem from our testing.