CMakeLists.txt: AOTRITON_INHERIT_SYSTEM_SITE_TRITON flag (#116)
Closes #114

## Motivation

Reusing an existing triton install allows building it separately from
aotriton.

This is useful for nixpkgs' packaging of aotriton as we would prefer to
build triton in a separate derivation and import it in aotriton.

## Technical Details

Added a new CMake flag that's off by default that enables the
`--system-site-packages` flag when creating the venv and checks an
existing triton can be imported instead of building it from the
submodule.

## Test Plan

Build with and without the flag enabled.

## Test Result

Works for both cases.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.