Alternative Wheels (#132)
# Overview

This PR allows to build different functionals with different triton wheels,
configured with a YAML-based configuration file.

The configuration file starts with a dictionary that specifies virtual
environments and the "alternative Triton wheels" that should be installed into
them.

The "rules" part of configuration file works like firewall rules: for each
rule, if all matches are successful, then the venv is used to compile this
functional. All members of Functional objects can be used in the matching rule.

## Limitations

This PR serves as an infrastructure to use tested Triton compiler for Navi3x
GPUs, but it does not fix the Navi3x accuracy issues found in 0.11b release by
itself, and does not handle the build of Triton wheels.

## Major Changes

* [build] AOTriton now requires `pyyaml`
* [build,codegen] Add cmake option `AOTRITON_ALT_TRITON_WHEEL_CONFIG_FILE` to specify
  the YAML configuration file.
* [docs] Add `docs/AltWheelExample.yaml` as an example of this configuration
  file.

## Minor Changes

* [build] Replace `git log -1 --format=%H` with `git rev-parse HEAD`
* [build] Now the build system throws error if SHA1 is not specified.
  + There are two ways to supply the AOTriton source GIT SHA1: detected by
    `git` automatically or supplied as env var `AOTRITON_CI_SUPPLIED_SHA1`
* [shim] Add property `.family` to class `Functional`
* [build] New Triton compiler will be installed under
  `<build_dir>/altvenvs/<venv_name>`, their cache directories are
  `<build_dir>/triton-cache.<venv_name>`
  * The triton cache directory for default triton wheel is renamed to `triton-cache.default`